requirements.txt
Django==3.2.14
djangorestframework==3.13.1
pymysql==1.0.2

